# programs_skills/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for competitor skills analysis",
        "title": "Competitor Skills API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/getCompetitorsSkills": {
            "get": {
                "description": "Runs the getCompetitorsSkills logic with default competitors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "competitors"
                ],
                "summary": "Get Competitors Skills (GET default)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Runs the getCompetitorsSkills logic with custom competitor list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "competitors"
                ],
                "summary": "Get Competitors Skills (POST with custom competitors)",
                "parameters": [
                    {
                        "description": "Competitor list",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CompetitorSkillsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CompetitorSkillsRequest": {
            "type": "object",
            "properties": {
                "competitors": {
                    "description": "List of competitors to analyze\nExample: [\"Amazon\", \"Google\", \"Microsoft\"]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}