# programs_skills/server.go
// @title        Competitor Skills API
// @version      1.0
// @description  API for competitor skills analysis
// @BasePath     /
// @schemes      http
// @servers      [{"url":"http://localhost:8009","description":"Local"},{"url":"http://3.86.69.179:8009","description":"Production"}]
package main

import (
    "github.com/gin-gonic/gin"
    "github.com/swaggo/gin-swagger"
    "github.com/swaggo/files"
    _ "programs_skills/docs" // docs is generated by Swag CLI
    "net/http"
    "log"
    "os"
    "encoding/json"
    "fmt"
    "os/exec"
    "bytes"
    "path/filepath"
    "strings"
)

// Request body for competitor skills analysis
// swagger:model CompetitorSkillsRequest
type CompetitorSkillsRequest struct {
    // List of competitors to analyze
    // Example: ["Amazon", "Google", "Microsoft"]
    Competitors []string `json:"competitors"`
}

// @Summary      Get Competitors Skills (POST with custom competitors)
// @Description  Runs the getCompetitorsSkills logic with custom competitor list
// @Tags         competitors
// @Accept       json
// @Produce      json
// @Param        request body CompetitorSkillsRequest true "Competitor list"
// @Success      200  {object}  map[string]interface{}
// @Failure      400  {object}  map[string]string
// @Failure      500  {object}  map[string]string
// @Router       /getCompetitorsSkills [post]
func getCompetitorsSkillsPostHandler(c *gin.Context) {
    var req CompetitorSkillsRequest
    if err := c.ShouldBindJSON(&req); err != nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid JSON", "details": err.Error()})
        return
    }
    
    if len(req.Competitors) == 0 {
        c.JSON(http.StatusBadRequest, gin.H{"error": "At least one competitor must be provided"})
        return
    }
    
    result, err := runCompetitorsSkillsWithCompetitors(req.Competitors)
    if err != nil {
        c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
        return
    }
    c.JSON(http.StatusOK, result)
}

// @Summary      Get Competitors Skills (GET default)
// @Description  Runs the getCompetitorsSkills logic with default competitors
// @Tags         competitors
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Failure      500  {object}  map[string]string
// @Router       /getCompetitorsSkills [get]
func getCompetitorsSkillsGetHandler(c *gin.Context) {
    result, err := runCompetitorsSkills()
    if err != nil {
        c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
        return
    }
    c.JSON(http.StatusOK, result)
}

func main() {
    r := gin.Default()
    
    // Swagger documentation
    r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
    
    // API endpoints
    r.GET("/getCompetitorsSkills", getCompetitorsSkillsGetHandler)
    r.POST("/getCompetitorsSkills", getCompetitorsSkillsPostHandler)
    
    // Health check endpoint
    r.GET("/health", func(c *gin.Context) {
        c.JSON(http.StatusOK, gin.H{"status": "healthy", "service": "competitor-skills-api"})
    })
    
    log.Println("Starting Competitor Skills API server with Swagger on :8009")
    log.Println("Swagger documentation available at: http://localhost:8009/swagger/index.html")
    log.Println("Server will be accessible on all interfaces (0.0.0.0:8009)")
    r.Run("0.0.0.0:8009")
}

// Helper function to run Go scripts and capture output
func runGoScript(scriptDir string) (interface{}, error) {
    return runGoScriptWithArgs(scriptDir, nil)
}

// Helper function to run Go scripts with competitor arguments
func runGoScriptWithArgs(scriptDir string, competitors []string) (interface{}, error) {
    currentDir, err := os.Getwd()
    if err != nil {
        return nil, fmt.Errorf("failed to get current directory: %w", err)
    }
    
    var scriptPath string
    scriptPath = filepath.Join(currentDir, scriptDir)
    if _, err := os.Stat(scriptPath); os.IsNotExist(err) {
        parentDir := filepath.Dir(currentDir)
        scriptPath = filepath.Join(parentDir, scriptDir)
        if _, err := os.Stat(scriptPath); os.IsNotExist(err) {
            return nil, fmt.Errorf("script directory %s not found in current directory %s or parent directory %s", scriptDir, currentDir, parentDir)
        }
    }
    
    originalDir, err := os.Getwd()
    if err != nil {
        return nil, fmt.Errorf("failed to get original directory: %w", err)
    }
    
    if err := os.Chdir(scriptPath); err != nil {
        return nil, fmt.Errorf("failed to change to script directory %s: %w", scriptPath, err)
    }
    defer os.Chdir(originalDir)
    
    args := []string{"run", "main.go"}
    if competitors != nil && len(competitors) > 0 {
        args = append(args, strings.Join(competitors, ","))
    }
    
    cmd := exec.Command("go", args...)
    var stdout, stderr bytes.Buffer
    cmd.Stdout = &stdout
    cmd.Stderr = &stderr
    
    if err := cmd.Run(); err != nil {
        return nil, fmt.Errorf("failed to run script in %s: %w\nStderr: %s", scriptPath, err, stderr.String())
    }
    
    output := strings.TrimSpace(stdout.String())
    
    // Try to parse as JSON first
    var jsonResult interface{}
    if err := json.Unmarshal([]byte(output), &jsonResult); err == nil {
        return jsonResult, nil
    }
    
    // If not JSON, return as string
    return map[string]string{"status": "success", "output": output}, nil
}

// Run competitors skills with default competitors
func runCompetitorsSkills() (interface{}, error) {
    log.Println("Running getCompetitorsSkills script with default competitors...")
    return runGoScript("getCompetitorsSkills")
}

// Run competitors skills with custom competitors
func runCompetitorsSkillsWithCompetitors(competitors []string) (interface{}, error) {
    log.Printf("Running getCompetitorsSkills script with custom competitors: %v", competitors)
    return runGoScriptWithArgs("getCompetitorsSkills", competitors)
} 